remove state logic from survey component
